requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.4.0
pydantic-settings==2.0.0
python-multipart==0.0.6
aiofiles==23.2.1
librosa==0.10.1
scipy==1.11.0
numpy==1.24.0
soundfile==0.12.1
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0

# Phonemization and Forced Alignment
phonemizer==3.2.1
praatio==6.0.0
textgrid==1.5
g2p-en==2.1.0
epitran==1.24

# Audio Analysis
praat-parselmouth==0.4.3
pyworld==0.3.4
crepe==0.0.12

# Scientific Computing
scikit-learn==1.3.0
dtw-python==1.3.0